backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
Flask-Caching==2.3.1
Flask-Limiter==4.1.1
celery==5.4.0
redis==5.2.1
python-dotenv==1.0.1
Werkzeug==3.1.3
pyotp==2.9.0
boto3==1.35.99
twilio==9.4.4
openai==1.66.3
pytest==8.3.5
